uniswapv3: added TickDataProvider and related utils functions (untested)
